Fix for TLS mode required in for example Azure Postgresql databases.
The TLS mode error message did not match with the error from the database, so TLS mode was never triggered. Fixed by adding 'SSL connection is required' to the error messages.

Because of the backoff algorithm the connection with TLS took about 30 seconds more. (30 seconds of failing to setup a non TLS connection before trying to connect with TLS). By accepting the sslmode=require parameter in the connection string this 30 seconds is skipped and TLS is immediately tried.